update: py tests

+++ src/py/mat3ra/mode/__init__.py
@@ -1,5 +1,5 @@
 import numpy as np
 
 
-def get_length(vec: np.ndarray) -> np.float_:
-    return np.linalg.norm(vec)
+def get_length(vec: np.ndarray) -> float:
+    return float(np.linalg.norm(vec))
